Start of updates to v4
Completely switches out the way that CSS files are managed.

Instead of overwriting core styles with a monolithic CSS file that gets enqueued, we hijack the enqueueing of the core CSS files and inject our own that have been modified.

This will be nice since after a core update it should be easy to just `diff` the files to see what needs to be updated.